Scene file for CoppeliaSim
The KUKA youBot is added to the simulator environment in CoppeliaSim. A perspective vision sensor is added to the base of the robot, set up with Persistence of Vision Ray Tracer (POV-Ray). A Floating view was also added so the sensor output is visible in real-time. Additionally, two cubics were added for better visibility.