The figure of the remote control
This code is also to be used in MATLAB.
The controller holds a button for validation connection to the simulation, a button to take a 
picture with the vision sensor, four sliders to manipulate the movement of the robotic arm, 
and a joystick to control the movement of the base. Each slider is configured with minimum 
and maximum values, based on the angles the arm can rotate. It is also able to display 
images taken by the vision sensor.